hello_world/requirements.txt
requests
jinja2
boto3
slackclient
